Fixed UI Buttons
Buttons now all work.
Difficulty buttons also changes color properly once clicked to make it more obvious which difficulty is selected.
