+++ Assets/Scripts/UI/UITesting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b67525bee368b41e8bf84f88fd90f3de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
